fold in official changes made on 12/11/07

+++ facebook-official/src/java/com/facebook/api/FacebookJsonRestClient.java
@@ -75,7 +75,7 @@ public class FacebookJsonRestClient extends ExtensibleClient<Object> {
     this._sessionKey = (String) d.get("session_key");
     Object uid = d.get("uid");
     try {
-      this._userId = (Integer) uid;
+      this._userId = ((Long) uid).intValue();
     } catch (ClassCastException cce) {
       this._userId = Integer.parseInt((String) uid);
     } 
@@ -104,9 +104,9 @@ public class FacebookJsonRestClient extends ExtensibleClient<Object> {
     if (json instanceof JSONObject) {
       JSONObject jsonObj = (JSONObject) json;
       if (jsonObj.containsKey("error_code")) {
-        Integer errorCode = (Integer) jsonObj.get("error_code");
+        Long errorCode = (Long) jsonObj.get("error_code");
         String message = (String) jsonObj.get("error_msg");
-        throw new FacebookException(errorCode, message);
+        throw new FacebookException(errorCode.intValue(), message);
       }
     }
     return json;
